feat: Add thread safe type WebviewMut
